requirements.txt
requests~=2.27.1
bs4~=0.0.1
beautifulsoup4~=4.10.0
prettytable~=3.0.0
colorgram.py